Uso de la camara nativa, PictureTools.java Modificacion del XML crear proyecto
