+++ app/src/main/res/layout/crear_proyecto.xml
@@ -133,6 +133,18 @@
         android:hint="@string/solicitar_al_area_de_ventas"
         android:inputType="textPersonName" />
 
+    <ImageView
+        android:id="@+id/imgFoto"
+        android:layout_width="200dp"
+        android:layout_height="200dp"
+        android:layout_marginTop="10dp"
+        android:contentDescription="Foto"
+        android:maxHeight="200dp"
+        android:maxWidth="200dp"
+        android:minHeight="200dp"
+        android:minWidth="200dp"
+        android:src="@android:drawable/ic_menu_camera"/>
+
     <Button
         android:id="@+id/Guardar"
         style="@style/ButtonStyle"
